kem: rename files for autogeneration of param parsing
