+++ .gitignore
@@ -99,6 +99,10 @@ providers/implementations/kdfs/scrypt.c
 providers/implementations/kdfs/sshkdf.c
 providers/implementations/kdfs/sskdf.c
 providers/implementations/kdfs/tls1_prf.c
+providers/implementations/kem/ec_kem.c
+providers/implementations/kem/ecx_kem.c
+providers/implementations/kem/ml_kem_kem.c
+providers/implementations/kem/rsa_kem.c
 providers/implementations/keymgmt/ml_dsa_kmgmt.c
 providers/implementations/keymgmt/ml_kem_kmgmt.c
 providers/implementations/keymgmt/mlx_kmgmt.c
