+++ build.info
@@ -62,6 +62,10 @@ DEPEND[]=include/openssl/asn1.h \
          providers/implementations/kdfs/sskdf.c \
          providers/implementations/kdfs/tls1_prf.c \
          providers/implementations/kdfs/x942kdf.c \
+         providers/implementations/kem/ec_kem.c \
+         providers/implementations/kem/ecx_kem.c \
+         providers/implementations/kem/ml_kem_kem.c \
+         providers/implementations/kem/rsa_kem.c \
          providers/implementations/keymgmt/ml_dsa_kmgmt.c \
          providers/implementations/keymgmt/ml_kem_kmgmt.c \
          providers/implementations/keymgmt/mlx_kmgmt.c \
@@ -120,6 +124,10 @@ DEPEND[providers/implementations/kdfs/argon2.c \
        providers/implementations/kdfs/sskdf.c \
        providers/implementations/kdfs/tls1_prf.c \
        providers/implementations/kdfs/x942kdf.c \
+       providers/implementations/kem/ec_kem.c \
+       providers/implementations/kem/ecx_kem.c \
+       providers/implementations/kem/ml_kem_kem.c \
+       providers/implementations/kem/rsa_kem.c \
        providers/implementations/keymgmt/ml_dsa_kmgmt.c \
        providers/implementations/keymgmt/ml_kem_kmgmt.c \
        providers/implementations/keymgmt/mlx_kmgmt.c \
@@ -159,6 +167,14 @@ GENERATE[providers/implementations/kdfs/tls1_prf.c]=\
     providers/implementations/kdfs/tls1_prf.c.in
 GENERATE[providers/implementations/kdfs/x942kdf.c]=\
     providers/implementations/kdfs/x942kdf.c.in
+GENERATE[providers/implementations/kem/ec_kem.c]=\
+    providers/implementations/kem/ec_kem.c.in
+GENERATE[providers/implementations/kem/ecx_kem.c]=\
+    providers/implementations/kem/ecx_kem.c.in
+GENERATE[providers/implementations/kem/ml_kem_kem.c]=\
+    providers/implementations/kem/ml_kem_kem.c.in
+GENERATE[providers/implementations/kem/rsa_kem.c]=\
+    providers/implementations/kem/rsa_kem.c.in
 GENERATE[providers/implementations/keymgmt/ml_dsa_kmgmt.c]=\
     providers/implementations/keymgmt/ml_dsa_kmgmt.c.in
 GENERATE[providers/implementations/keymgmt/ml_kem_kmgmt.c]=\
