+++ providers/implementations/kem/ec_kem.c.in
@@ -34,7 +34,7 @@
 #include "internal/hpke_util.h"
 #include "crypto/ec.h"
 #include "prov/ecx.h"
-#include "eckem.h"
+#include "prov/eckem.h"
 
 typedef struct {
     EC_KEY *recipient_key;
