+++ providers/implementations/kem/ecx_kem.c.in
@@ -34,7 +34,7 @@
 #include "crypto/ecx.h"
 #include <openssl/hpke.h>
 #include "internal/hpke_util.h"
-#include "eckem.h"
+#include "prov/eckem.h"
 
 #define MAX_ECX_KEYLEN X448_KEYLEN
